refactor(loom): tighten runtime module boundaries
