+++ crates/loom/src/lib.rs
@@ -12,9 +12,9 @@ pub mod auth;
 pub mod automation;
 pub mod backend;
 pub mod builtins;
-pub mod changes;
+pub(crate) mod changes;
 pub mod chat;
-pub mod chatlog;
+pub(crate) mod chatlog;
 pub mod client;
 pub mod client_context;
 pub mod custom_agents;
@@ -26,9 +26,9 @@ pub mod github;
 pub mod github_app;
 pub mod github_manifest;
 pub mod github_trigger;
-pub mod history;
+pub(crate) mod history;
 pub mod ide;
-pub mod launch;
+pub(crate) mod launch;
 pub mod launch_gate;
 pub mod logs;
 pub mod loom_config;
@@ -44,20 +44,39 @@ pub mod runs;
 pub(crate) mod runtime;
 pub mod server;
 pub mod session;
-pub mod session_layout;
+pub(crate) mod session_layout;
 pub mod session_manager;
-pub mod setup;
+pub(crate) mod setup;
 pub mod shell;
-pub mod slack;
-pub mod tasks;
-pub mod terminal;
+pub(crate) mod slack;
+pub(crate) mod tasks;
+pub(crate) mod terminal;
 pub mod user_token;
 pub mod watch;
 pub mod web;
 
-pub use web::AppState;
+/// Shared process state consumed by Loom's runtime services and REST adapter.
+#[derive(Clone)]
+pub struct AppState {
+    pub db: db::Db,
+    pub bus: weaver_core::events::EventBus,
+    /// host:port the server is bound to, used to build child-process env.
+    pub addr: String,
+    /// Per-session embedded code-server lifecycle + reverse-proxy registry.
+    pub ide: std::sync::Arc<ide::IdeManager>,
+    /// The inbound GitHub trigger: its GitHub gateway (the `gh`-backed default)
+    /// and per-repo rate limiter. Shared across requests; a test swaps in a fake
+    /// gateway via [`github_trigger::GithubTrigger::with_gateway`].
+    pub trigger: std::sync::Arc<github_trigger::GithubTrigger>,
+    /// The registry of live ACP session tasks — the conversation routes drive
+    /// sessions through it and subscribe to its SSE stream.
+    pub acp: acp::AcpRegistry,
+    /// Namespaced repository provisioning, capped-profile admission, and
+    /// per-session Scratch mutation locks.
+    pub launch_gate: launch_gate::RepoLaunchGate,
+}
 
-// Re-export weaver-core modules at the same paths so the orchestrator code can
-// continue to use short `crate::events`/`crate::git`/etc. references.
-pub use weaver_core::db::Db;
-pub use weaver_core::{branch, config, events, git, issue};
+// Crate-local aliases keep Loom implementation imports short without exposing
+// weaver-core's storage and domain modules as part of Loom's public API.
+pub(crate) use weaver_core::db::Db;
+pub(crate) use weaver_core::{branch, config, events, git};
