+++ crates/loom/tests/repo_setup.rs
@@ -11,13 +11,13 @@ use std::path::{Path, PathBuf};
 use std::process::Command;
 use std::time::Duration;
 
-use loom::events::EventBus;
 use loom::session as session_mod;
-use loom::web::AppState;
+use loom::AppState;
 use loom::{agent_env, client, db, repo, repo_env, server};
 use serde_json::{json, Value};
 use serial_test::serial;
 use tokio::net::TcpListener;
+use weaver_core::events::EventBus;
 
 fn sh(dir: &Path, program: &str, args: &[&str]) {
     let status = Command::new(program)
